+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0](https://github.com/owkin/connectlib/releases/tag/0.15.0) - 2022-05-31
+
+### Added
+
 - feat: meaningful name for algo (#175). You can use the `_algo_name` parameter to set a custom algo name for the registration. By default, it is set to `method-name_class-name`.
 
   ```py